# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FocustApps Fenrir Test Automation",
    "description": "Fenrir is a test automation tool for web applications.",
    "version": "0.1"
  },
  "paths": {
...
    "/v1/api/accounts/": {
      "post": {
        "tags": [
          "accounts-api"
        ],
        "summary": "Create Account",
        "description": "Create a new account.\n\nAny authenticated user can create an account - they automatically become the owner.\nA user-account association is also created with 'owner' role and marked as primary.\n\nArgs:\n    request: FastAPI request object (for IP tracking)\n    body: Account creation data\n    current_user: JWT token payload\n\nReturns:\n    AccountResponse: Created account data\n\nRaises:\n    HTTPException: 400 if account name already exists",
        "operationId": "create_account_v1_api_accounts__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAccountRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "accounts-api"
        ],
        "summary": "List User Accounts",
        "description": "List all accounts accessible to the current user.\n\nArgs:\n    include_inactive: Whether to include inactive accounts\n    current_user: JWT token payload\n\nReturns:\n    List[UserAccountsResponse]: List of user's accounts with roles\n\nNotes:\n    - Super admins see all accounts\n    - Regular users see only accounts they're associated with\n    - Results include role and primary account flag",
        "operationId": "list_user_accounts_v1_api_accounts__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "include_inactive",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Inactive"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserAccountsResponse"
                  },
                  "title": "Response List User Accounts V1 Api Accounts  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AccountResponse": {
        "properties": {
          "account_id": {
            "type": "string",
            "title": "Account Id"
          },
          "account_name": {
            "type": "string",
            "title": "Account Name"
          },
          "owner_user_id": {
            "type": "string",
            "title": "Owner User Id"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "logo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo Url"
          },
          "primary_contact": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Contact"
          },
          "subscription_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subscription Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "account_id",
          "account_name",
          "owner_user_id",
          "is_active",
          "created_at"
        ],
        "title": "AccountResponse",
        "description": "Response model for basic account data."
      },
...
      "CreateAccountRequest": {
        "properties": {
          "account_name": {
            "type": "string",
            "title": "Account Name"
          }
        },
        "type": "object",
        "required": [
          "account_name"
        ],
        "title": "CreateAccountRequest",
        "description": "Request model for creating a new account."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UserAccountsResponse": {
        "properties": {
          "account_id": {
            "type": "string",
            "title": "Account Id"
          },
          "account_name": {
            "type": "string",
            "title": "Account Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "is_primary": {
            "type": "boolean",
            "title": "Is Primary"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "account_id",
          "account_name",
          "role",
          "is_primary",
          "is_active"
        ],
        "title": "UserAccountsResponse",
        "description": "Response model for listing user's accessible accounts."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}